LectureBank 관련 기능 변경

+++ src/main/java/in/hangang/serviceImpl/LectureBankServiceImpl.java
@@ -2,6 +2,7 @@ package in.hangang.serviceImpl;
 
 import in.hangang.domain.*;
 import in.hangang.enums.ErrorMessage;
+import in.hangang.enums.Point;
 import in.hangang.exception.RequestInputException;
 import in.hangang.mapper.LectureBankMapper;
 import in.hangang.mapper.UserMapper;
@@ -9,16 +10,16 @@ import in.hangang.service.LectureBankService;
 import in.hangang.service.UserService;
 import in.hangang.util.S3Util;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.io.InputStreamSource;
-import org.springframework.core.io.Resource;
-import org.springframework.core.io.UrlResource;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
+
+import static in.hangang.enums.ErrorMessage.INVALID_ACCESS_EXCEPTION;
 
 @Service
 public class LectureBankServiceImpl implements LectureBankService {
@@ -62,7 +63,7 @@ public class LectureBankServiceImpl implements LectureBankService {
     public LectureBank getLectureBank(Long id) throws Exception{
         LectureBank lectureBank = lectureBankMapper.getLectureBank(id);
         if(lectureBank == null)
-            throw new RequestInputException(ErrorMessage.INVALID_ACCESS_EXCEPTION);
+            throw new RequestInputException(INVALID_ACCESS_EXCEPTION);
         User user = userMapper.getMe(lectureBank.getUser_id());
         lectureBank.setUser(user);
 
@@ -92,26 +93,29 @@ public class LectureBankServiceImpl implements LectureBankService {
         if(lectureBank.getId() == null) throw new RequestInputException(ErrorMessage.REQUEST_INVALID_EXCEPTION);
         LectureBank lb = lectureBankMapper.getLectureBank(lectureBank.getId());
 
-        Long lecture_id; String title, content; Integer point_price;
-        if(lectureBank.getLecture_id() == null) lecture_id = lb.getLecture_id();
-        else lecture_id = lectureBank.getLecture_id();
-        if(lectureBank.getTitle() == null) title = lb.getTitle();
-        else title = lectureBank.getTitle();
-        if(lectureBank.getContent() == null) content = lb.getContent();
-        else content = lectureBank.getContent();
-        if(lectureBank.getPoint_price() == null) point_price = lb.getPoint_price();
-        else point_price = lectureBank.getPoint_price();
+        Long lecture_id=0L; String title="", content=""; Integer point_price=0;
 
-        lectureBankMapper.setLectureBank(lectureBank.getId(), lecture_id, title, content, point_price);
+        if(lectureBank.getLecture_id() != null) lecture_id = lectureBank.getLecture_id();
+        else if(lb.getLecture_id() != null) lecture_id = lb.getLecture_id();
+        if(lectureBank.getTitle() != null) title = lectureBank.getTitle();
+        else if(lb.getTitle() != null) title = lb.getTitle();
+        if(lectureBank.getContent() != null) content = lectureBank.getContent();
+        else if(lb.getContent() != null) content = lb.getContent();
+        if(lectureBank.getPoint_price() != null) point_price = lectureBank.getPoint_price();
+        else if(lb.getPoint_price() != null) point_price = lb.getPoint_price();
+
+        Long semester_date_id = lectureBankMapper.getLatestSemesterID(); String date;
+        if(lectureBank.getSemester_date() == null)  date = lb.getSemester_date();
+        else date = lectureBank.getSemester_date();
+        semester_date_id = lectureBankMapper.getSemesterID(date);
+        lectureBankMapper.setLectureBank(lectureBank.getId(), lecture_id, title, content, point_price, semester_date_id);
 
         List<String> categoryList = lectureBank.getCategory();
         if(categoryList != null){
-            //TODO 이해가 안되는 부분 왜 강의자료를 삽입할 때 댓글을 지우는지? 이 카테고리의 id는 댓글ID과 무슨 상관인지?
-            //TODO 수정에서도 이 함수를 사용하던데 이 때 카테고리를 지워야하는 거 아닌가요?
             //lectueMapper add to lecturebank _ Category : category list while empty
             List<Long> idList = lectureBankMapper.getCategoryIdList(lectureBank.getId());
             for(Long id : idList){
-                lectureBankMapper.deleteComment(id);
+                lectureBankMapper.deleteCategory(id);
             }
             //TODO 자바에서 디비를 반복적으로 접근하는건 매우 비효율 적입니다 mybatis의 for each를 사용해주세요
             for (String s : categoryList) {
@@ -131,21 +135,22 @@ public class LectureBankServiceImpl implements LectureBankService {
             //TODO 자바에서 디비를 반복적으로 접근하는건 매우 비효율 적입니다 mybatis의 for each를 사용해주세요
             lectureBankMapper.setFileAvailable(i,1);
         }
-        //TODO 어차피 인터셉터에서 걸립니다 Controller에서 @Auth 어노테이션을 사용했으므로 해당 코드는 필요가 없습니다.
+        //TODO zeplin: 작성시 +50point -> POINT
         User user = userService.getLoginUser();
-        if (user==null) throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
-
-        //zeplin: 작성시 +50point
         lectureBankMapper.setPoint(user.getId(), 50);
     }
 
     @Override
+    @Transactional
     public void deleteLectureBank(Long id) throws Exception{
-
         //TODO Trancation 처리를 한 뒤 해당 강의자료가 지워지면 댓글, 파일, 카테고리 등 연관된 부분도 전부 지워지면 좋을 것 같습니다
         //TODO soft delete를 적용해주면 나중에 더 좋을 것 같습니다!
+
+        //user_point_history
+        // point_type
         User user = userService.getLoginUser();
         if (user==null) throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
+
         lectureBankMapper.deleteLectureBank(id, user.getId());
     }
 
@@ -157,55 +162,56 @@ public class LectureBankServiceImpl implements LectureBankService {
 
     @Override
     public void addComment(Long lecture_bank_id, String comments) throws Exception{
-        User user = userService.getLoginUser();
-        if (user==null) throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
-        lectureBankMapper.addComment(user.getId(), lecture_bank_id, comments);
+        lectureBankMapper.addComment(userService.getLoginUser().getId(), lecture_bank_id, comments);
     }
 
     @Override
     public void setComment(Long lecture_bank_comment_id, String comments) throws Exception{
-        lectureBankMapper.setComment(lecture_bank_comment_id, comments);
+        if(checkCommentWriter(lecture_bank_comment_id))
+            lectureBankMapper.setComment(lecture_bank_comment_id, comments);
+        else
+            throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
     }
 
     @Override
     public void deleteComment(Long lecture_bank_comment_id) throws Exception{
-        lectureBankMapper.deleteComment(lecture_bank_comment_id);
+        if(checkCommentWriter(lecture_bank_comment_id))
+            lectureBankMapper.deleteComment(lecture_bank_comment_id);
+        else
+            throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
+    }
+
+    @Override
+    public Boolean checkCommentWriter(Long lecture_bank_comment_id)throws Exception{
+        Long userID = userService.getLoginUser().getId();
+        Long writerID = lectureBankMapper.getCommentWriterId(lecture_bank_comment_id);
+
+        return userID.equals(writerID);
     }
 
     //purchase------------------------------------------------------------------------------------
     @Override
     public Boolean checkPurchase(Long lecture_bank_id) throws Exception {
         User user = userService.getLoginUser();
-        if (user==null) throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
         Integer purchase = lectureBankMapper.checkPurchased(user.getId(), lecture_bank_id);
-        if(purchase == null){
-            return false;
-        }
-        if(purchase.intValue() == 1){
-            return true;
-        }else{
-            return false;
-        }
+        return purchase != null;
     }
     @Override
+    @Transactional
     public void purchase(Long lecture_bank_id) throws Exception{
-        User user = userService.getLoginUser();
-        if (user==null) throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
+        if(checkPurchase(lecture_bank_id)) throw new RequestInputException(ErrorMessage.INVALID_ACCESS_EXCEPTION);
+
+        Long userID = userService.getLoginUser().getId();
         LectureBank lectureBank = getLectureBank(lecture_bank_id);
-        Integer purchase = lectureBankMapper.checkPurchased(user.getId(), lecture_bank_id);
-
         Integer point_price = lectureBank.getPoint_price();
-        if(purchase == null){
-            // purchased table에 추가
-            lectureBankMapper.purchaseInsert(user.getId(), lecture_bank_id);
-        }
 
-        if(purchase.intValue()==0){
-            // 구매한 유저 포인트-- 자료주인은 포인트++
-            lectureBankMapper.setPoint(user.getId(), -point_price);
-            lectureBankMapper.setPoint(lectureBank.getUser_id(), point_price);
+        // 구매한 유저 포인트-- 자료주인은 포인트++
+        lectureBankMapper.purchaseInsert(userID, lecture_bank_id);
+        lectureBankMapper.setPoint(userID, -point_price);
+        lectureBankMapper.addPointHistory(userID,-point_price, Point.LECTURE_PURCHASE.getTypeId());
+        lectureBankMapper.setPoint(lectureBank.getUser_id(), point_price);
+        lectureBankMapper.addPointHistory(userID,point_price,Point.LECTURE_SELL.getTypeId());
 
-        }
 
     }
 
@@ -213,33 +219,26 @@ public class LectureBankServiceImpl implements LectureBankService {
     //hits------------------------------------------------------------------------------------
     @Override
     public Boolean checkHits(Long lecture_bank_id) throws  Exception {
-        User user = userService.getLoginUser();
-        if (user==null) throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
-        Integer hits = lectureBankMapper.checkHits(user.getId(), lecture_bank_id);
-        if(hits == null){
-            return false;
-        }
-        if(hits.intValue() ==1){
-            return true;
-        }else{
-            return false;
-        }
+        Long userID = userService.getLoginUser().getId();
+        Integer hits = lectureBankMapper.checkHits(userID, lecture_bank_id);
+        return hits != null;
     }
 
     @Override
+    @Transactional
     public void pushHit(Long lecture_bank_id) throws Exception{
-        User user = userService.getLoginUser();
-        if (user==null) throw new RequestInputException(ErrorMessage.INVALID_USER_EXCEPTION);
-        Integer hits = lectureBankMapper.checkHits(user.getId(), lecture_bank_id);
+        Long userID = userService.getLoginUser().getId();
+
+        Integer hits = lectureBankMapper.checkHits(userID, lecture_bank_id);
+
         if(hits == null){
-            //insert mapper with hit 1
-            lectureBankMapper.hitInsert(user.getId(), lecture_bank_id);
+            lectureBankMapper.hitInsert(userID, lecture_bank_id);
             lectureBankMapper.addHit_lecture_bank(lecture_bank_id);
         } else if(hits.intValue()==1){
-            lectureBankMapper.subHit(user.getId(), lecture_bank_id);
+            lectureBankMapper.subHit(userID, lecture_bank_id);
             lectureBankMapper.subHit_lecture_bank(lecture_bank_id);
         }else{
-            lectureBankMapper.addHit(user.getId(), lecture_bank_id);
+            lectureBankMapper.addHit(userID, lecture_bank_id);
             lectureBankMapper.addHit_lecture_bank(lecture_bank_id);
         }
     }
@@ -277,16 +276,36 @@ public class LectureBankServiceImpl implements LectureBankService {
         return lectureBankMapper.getFileList(lecture_bank_id);
     }
 
+    /*
     @Override
     public org.springframework.core.io.Resource getprivateObject(Long id) throws Exception{
-        //TODO 다운로드 받을 유저의 권한체크 필요할 것 같습니다 1. 게시자인지/ 2. 구매자인지
         String objectKey = lectureBankMapper.getUrl(id);
         URL url = s3Util.getPrivateObjectURL(objectKey);
-
-        //TODO 어 그런데 생각해보니까 굳이 뻘짓할 필요없이 싹다 PRIVATE로 하고 권한 체크한뒤  그냥 파일 자체를 서버에서 S3.GET해서 던져주면 됐던거아닌가요?
         org.springframework.core.io.Resource resource = new UrlResource(url);
         return resource;
     }
+
+     */
+
+    @Override
+    public String getObjectUrl(Long id) throws Exception{
+        //TODO 다운로드 받을 유저의 권한체크 필요할 것 같습니다 1. 게시자인지/ 2. 구매자인지
+
+        Long user_id = userService.getLoginUser().getId();
+        Long lecturebank_id = lectureBankMapper.getLectureBankId_file(id);
+        Long writer = lectureBankMapper.getWriterId(lecturebank_id);
+        if(checkPurchase(lecturebank_id) || (writer.equals(user_id))){
+            String objectKey = lectureBankMapper.getUrl(id);
+            URL url = s3Util.getPrivateObjectURL(objectKey);
+            return url.toString();
+        }else{
+
+            return null;
+        }
+
+    }
+
+
 
     @Override
     public void cancelUpload(Long id) throws Exception{
